refactor: organize client scripts
