+++ console.html
@@ -185,8 +185,8 @@
         </div>
     </div>
 
-    <script src="js/webadb.js"></script>
-    <script src="js/auto-scale.js"></script>
+    <script src="js/adb/webadb.js"></script>
+    <script src="js/ui/auto-scale.js"></script>
     <script src="js/apps.js" type="module"></script>
 </body>
 </html>
