+++ docs/css/styles.css
@@ -884,8 +884,8 @@ body.scaled {
     border: 2px solid var(--border-color);
     border-radius: 0.5rem;
     padding: 1rem;
-    min-height: 400px;
-    max-height: 600px;
+    min-height: 300px;
+    max-height: 500px;
     overflow-y: auto;
     overflow-x: auto;
     font-family: 'Courier New', monospace;
