+++ docs/js/adb/apk-installer.js
