+++ docs/js/ui/cards.js
@@ -47,3 +47,4 @@ export function renderKits(apks, { onInstall } = {}) {
         loop: true
     });
 }
+
