+++ js/adb/apk-installer.js
