+++ template.html
@@ -196,9 +196,9 @@
         </div>
     </div>
 
-    <script src="js/webadb.js"></script>
+    <script src="js/adb/webadb.js"></script>
     <script src="https://cdn.jsdelivr.net/npm/swiper@11/swiper-bundle.min.js"></script>
-    <script src="js/auto-scale.js"></script>
+    <script src="js/ui/auto-scale.js"></script>
     <script src="js/apps.js" type="module"></script>
 </body>
 </html>
